HotelService/HotelService/src/main/java/com/jvd/hotel/service/impl/HotelServiceImpl.java
package com.jvd.hotel.service.impl;

import com.jvd.hotel.entities.Hotel;
import com.jvd.hotel.exception.ResourceNotFoundException;
import com.jvd.hotel.repositories.HotelRepository;
import com.jvd.hotel.service.HotelService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.UUID;
@Service
public class HotelServiceImpl implements HotelService {
    @Autowired
    private HotelRepository hotelRepository;
    @Override
    public Hotel create(Hotel hotel) {
        String randomId = UUID.randomUUID().toString();
        hotel.setId(randomId);
        return hotelRepository.save(hotel);
    }

    @Override
    public List<Hotel> getAll() {
        return hotelRepository.findAll();
    }

    @Override
    public Hotel get(String id) {
        return hotelRepository.findById(id)
                .orElseThrow(()-> new ResourceNotFoundException("Hotel that given id is not found"));
    }
}
